Fix comment mappings in commit
https://github.com/wildfly/wildfly/commit/d7675fb0b19d3d22978e79954f441eeefd74a3b2
For comments extracted to new methods, there might be multiple callsites.
Some of them are part of new functionality, simply reusing the extracted method.
The update finds the callsite corresponding to the original extracted functionality.
The callsite affects the parentMapping, which in turn determines the deleted comments of the parentMapper that should processed and matched with comments inside the extracted method.